Reorganize client code
