+++ src/app/auth/components/AuthCard.tsx
