+++ src/app/auth/signup/page.tsx
@@ -1,7 +1,7 @@
 import { authOptions } from '@/lib/auth/options';
 import { getServerSession } from 'next-auth';
 import { redirect } from 'next/navigation';
-import SignupForm from '../../_modules/auth/components/SignupForm';
+import SignupForm from '../components/SignupForm';
 
 const SignupPage = async () => {
   const session = await getServerSession(authOptions);
