+++ src/app/common/components/Logo.tsx
