+++ src/app/common/components/icons/EmailIcon.tsx
