+++ src/app/common/components/icons/LogoIcon.tsx
