+++ src/app/common/components/icons/PasswordIcon.tsx
