+++ src/app/page.tsx
@@ -1,6 +1,6 @@
 import { getServerSession } from "next-auth";
 import { redirect } from "next/navigation";
-import LogoutButton from "./_modules/auth/components/LogoutButton";
+import LogoutButton from "./auth/components/LogoutButton";
 import { authOptions } from "@/lib/auth/options";
 
 const HomePage = async () => {
